// frontend/src/types/tGR8Game.ts
export type tGR8Game = {
  id: string;
  walletId: string;
  productType: string;
  translationKey: string;
  gameProvider: string;
  channel: string;
  dealerLanguage: string;
  imageUrl: string | null;
  isDemoModeAvailable: boolean;
  isFreeSpinsAvailable: boolean;
  isActive: boolean;
  isDeleted: boolean;
  img_custom: string | null;
};